Widen "order" columns to double precision
Drag-to-reorder assigns the midpoint between two neighbours so a move rewrites
one row instead of renumbering the list. Midpoints are fractional, but the
columns were left as integer when that landed, so the first drag failed with
'invalid input syntax for type integer: "0.5"'.

Widens tasks/habits/goals."order", and the local variable in quick_add that
computes an append position. Fixes schema.sql too so a fresh install does not
inherit the same defect. weekday stays integer — it genuinely is one.

+++ supabase/migrate-002-recurrence-subtasks-shortcuts.sql
@@ -90,7 +90,7 @@ declare
   v_user  uuid;
   v_id    text;
   v_date  text;
-  v_order integer;
+  v_order double precision;
 begin
   if p_text is null or length(btrim(p_text)) = 0 then
     raise exception 'empty task';
